Finish Product middleware sturcture refactor
